+++ lib/src/hive_local_storage.dart
@@ -2,30 +2,23 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
+// ignore: depend_on_referenced_packages
+import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:hive_flutter/hive_flutter.dart';
 
 import 'jwt_decoder.dart';
 import 'session.dart';
-import 'storage_keys.dart';
 
 /// {@template local_storage}
 /// A wrapper class for session and cache box uses [Hive]
 /// {@endtemplate}
 class LocalStorage {
-  ///{@macro local_storage}
-  LocalStorage._();
-
-  /// static instance
-  static LocalStorage get _instance => LocalStorage._();
-
-  /// factory constructors are used for Riverpod only
-  /// do not user for normal purpose
-  /// Example:
+  /// Example: use with Riverpod
   ///
   /// ```dart
-  /// final localStorageProvider = Provider<LocalStorage>((ref)=>LocalStorage());
+  /// final localStorageProvider = Provider<LocalStorage>((ref)=>throw UnImplementedError());
   /// ```
   ///
   /// in main function
@@ -51,9 +44,17 @@ class LocalStorage {
 
   ///
   ///
-  factory LocalStorage() {
-    return _instance;
-  }
+  ///{@macro local_storage}
+  LocalStorage._();
+
+  /// session key
+  static const String sessionKey = '__JWT_SESSION_KEY__';
+
+  /// cache key
+  static const String cacheKey = '__CACHE_KEY__';
+
+  /// encryption key
+  static const String encryptionBoxKey = '__ENCRYPTION_KEY__';
 
   /// [FlutterSecureStorage] storage for storing encryption key
   static late FlutterSecureStorage _storage;
@@ -61,19 +62,16 @@ class LocalStorage {
   /// [Box] session box
   static late Box<Session> _sessionBox;
 
-  /// [Box] encrypted box
-  static late Box<dynamic> _encryptedBox;
-
   /// [Box] _cacheBox
   static late Box<dynamic> _cacheBox;
 
-  /// init
   /// initialize the dependencies
   /// register the adapters
   /// open the boxes
   /// returns [LocalStorage] instance
-  static Future<LocalStorage> getInstance(
-      [List<TypeAdapter<HiveObject>>? adapters]) async {
+  static Future<LocalStorage> getInstance({
+    List<TypeAdapter<HiveObject>>? adapters,
+  }) async {
     WidgetsFlutterBinding.ensureInitialized();
     if (adapters != null && adapters.isNotEmpty) {
       for (final adapter in adapters) {
@@ -87,27 +85,19 @@ class LocalStorage {
     );
     final encryptionCipher = await _encryptionKey;
     await Hive.initFlutter();
-    _sessionBox = await Hive.openBox<Session>(
-      StorageKeys.sessionKey,
-      encryptionCipher: encryptionCipher,
-    );
-    _encryptedBox = await Hive.openBox<dynamic>(
-      StorageKeys.encryptedBoxKey,
-      encryptionCipher: encryptionCipher,
-    );
-    _cacheBox = await Hive.openBox<dynamic>(StorageKeys.cacheKey);
-    return _instance;
+    _sessionBox = await Hive.openBox<Session>(sessionKey);
+    _cacheBox = await Hive.openBox(cacheKey, encryptionCipher: encryptionCipher);
+    return LocalStorage._();
   }
 
   /// HiveAesCipher encryptionKey
   /// encryption key to secure session box
   static Future<HiveAesCipher> get _encryptionKey async {
     late Uint8List encryptionKey;
-    var keyString = await _storage.read(key: StorageKeys.encryptionKey);
+    var keyString = await _storage.read(key: encryptionBoxKey);
     if (keyString == null) {
       final key = Hive.generateSecureKey();
-      await _storage.write(
-          key: StorageKeys.encryptionKey, value: base64UrlEncode(key));
+      await _storage.write(key: encryptionBoxKey, value: base64UrlEncode(key));
       encryptionKey = Uint8List.fromList(key);
     } else {
       encryptionKey = base64Url.decode(keyString);
@@ -115,10 +105,88 @@ class LocalStorage {
     return HiveAesCipher(encryptionKey);
   }
 
+  /// open custom box
+  Future<void> openCustomBox<T extends HiveObject>({required String boxName, required int typeId}) async {
+    if (!Hive.isAdapterRegistered(typeId)) {
+      throw Exception('Please register adapter for $T before opening box.');
+    }
+    await Hive.openBox<T>(boxName, encryptionCipher: await _encryptionKey);
+  }
+
+  /// `getCustomList`
+  /// get data from custom box
+  List<T> getCustomList<T extends HiveObject>({
+    required String boxName,
+  }) {
+    if (Hive.isBoxOpen(boxName)) {
+      final box = Hive.box<T>(boxName);
+      return box.values.toList();
+    } else {
+      throw Exception('Please `openCustomBox` before accessing it');
+    }
+  }
+
+  /// `add`
+  /// add data to custom box
+  Future<void> add<T extends HiveObject>({
+    required String boxName,
+    required T value,
+  }) async {
+    if (Hive.isBoxOpen(boxName)) {
+      final box = Hive.box<T>(boxName);
+      await box.add(value);
+    } else {
+      throw Exception('Please `openCustomBox` before accessing it');
+    }
+  }
+
+  /// `addAll`
+  /// add multiple data to custom box
+  Future<void> addAll<T extends HiveObject>({
+    required String boxName,
+    required List<T> values,
+  }) async {
+    if (Hive.isBoxOpen(boxName)) {
+      final box = Hive.box<T>(boxName);
+      await box.addAll(values);
+    } else {
+      throw Exception('Please `openCustomBox` before accessing it');
+    }
+  }
+
+  /// `update`
+  /// update item from data
+  Future<void> update<T extends HiveObject>({required String boxName, required T value}) async {
+    if (Hive.isBoxOpen(boxName)) {
+      final box = Hive.box<T>(boxName);
+      final data = box.values.firstWhereOrNull((element) => element == value);
+      if (data != null) {
+        await data.delete();
+        await box.add(value);
+      } else {
+        await box.add(value);
+      }
+    } else {
+      throw Exception('Please `openCustomBox` before accessing it');
+    }
+  }
+
+  /// `delete`
+  /// delete item from data
+  Future<void> delete<T extends HiveObject>({required String boxName, required T value}) async {
+    if (Hive.isBoxOpen(boxName)) {
+      final box = Hive.box<T>(boxName);
+      final data = box.values.firstWhereOrNull((element) => element == value);
+      await data?.delete();
+    } else {
+      throw Exception('Please `openCustomBox` before accessing it');
+    }
+  }
+
   /// `getSession`
   /// get [Session] from the box
   Session? getSession() {
-    return _sessionBox.isNotEmpty ? _sessionBox.getAt(0) : null;
+    return _sessionBox.getAt(0);
   }
 
   /// `hasSession`
@@ -155,12 +223,19 @@ class LocalStorage {
   T? get<T>({
     required String key,
     T? defaultValue,
-    bool useEncryption = false,
   }) {
-    if (useEncryption) {
-      return _encryptedBox.get(key, defaultValue: defaultValue);
-    } else {
-      return _cacheBox.get(key, defaultValue: defaultValue);
+    return _cacheBox.get(key, defaultValue: defaultValue);
+  }
+
+  /// get list data
+  List<T> getList<T>({required String key}) {
+    try {
+      final String encodedData = _cacheBox.get(key, defaultValue: '');
+      if (encodedData.isEmpty) return [];
+      final decodedData = jsonDecode(_cacheBox.get(key));
+      return List<T>.from(jsonDecode(decodedData));
+    } catch (_) {
+      return [];
     }
   }
 
@@ -169,65 +244,45 @@ class LocalStorage {
   Future<void> put<T>({
     required String key,
     required T value,
-    bool useEncryption = false,
   }) async {
-    if (useEncryption) {
-      return _encryptedBox.put(key, value);
-    } else {
-      return _cacheBox.put(key, value);
-    }
+    return _cacheBox.put(key, value);
+  }
+
+  /// save list of data
+  Future<void> putList<T>({
+    required String key,
+    required List<T> value,
+  }) async {
+    final encodedData = jsonEncode(value);
+    return _cacheBox.put(key, encodedData);
   }
 
   /// save
   /// puts value in box with [key]
   Future<void> putAll({
     required Map<String, dynamic> entries,
-    bool useEncryption = false,
   }) async {
-    if (useEncryption) {
-      return _encryptedBox.putAll(entries);
-    } else {
-      return _cacheBox.putAll(entries);
-    }
+    return _cacheBox.putAll(entries);
   }
 
   /// remove
   /// removes value from box registered with [key]
-  Future<void> remove({
-    required String key,
-    bool useEncryption = false,
-  }) async {
-    if (useEncryption) {
-      return _encryptedBox.delete(key);
-    } else {
-      return _cacheBox.delete(key);
-    }
+  Future<void> remove({required String key}) async {
+    return _cacheBox.delete(key);
   }
 
   /// clear
   /// clear all values from box
-  Future<int> clear({bool useEncryption = false}) async {
-    if (useEncryption) {
-      return _encryptedBox.clear();
-    } else {
-      return _cacheBox.clear();
-    }
+  Future<int> clear() async {
+    return _cacheBox.clear();
   }
 
   /// close all the opened box
   Future<void> closeAll() async {
-    await Future.wait([
-      _sessionBox.close(),
-      _encryptedBox.close(),
-      _cacheBox.close(),
-    ]);
+    await _sessionBox.close();
+    await _cacheBox.close();
   }
 
   /// convert box to map
-  Map<String, Map<String, dynamic>?> toCacheMap() =>
-      Map.unmodifiable(_cacheBox.toMap());
-
-  /// convert box to map
-  Map<String, Map<String, dynamic>?> toEncryptedMap() =>
-      Map.unmodifiable(_encryptedBox.toMap());
+  Map<String, Map<String, dynamic>?> toCacheMap() => Map.unmodifiable(_cacheBox.toMap());
 }