+++ lib/src/storage_keys.dart
@@ -1,6 +1,0 @@
-abstract class StorageKeys {
-  static const String sessionKey = '__JWT_SESSION_KEY__';
-  static const String encryptedBoxKey = '__ENCRYPTED_CACHE_KEY__';
-  static const String cacheKey = '__CACHE_KEY__';
-  static const String encryptionKey = '__ENCRYPTION_KEY__';
-}